:artificial_satellite: Avoiding negative value parsing inconsistency
You might have had inputs with `my_command - true 1` where the minus would have applied to the `1`. Any minus tokens in the line would have also been ignored or not applied if a number is not found afterwards.

To catch these issues, now a minus token just gets merged with the next token if it's a number, and we just let the `string_to_int/string_to_float64` procedures handle it.