:feather: Adding access to command data by name, also to the Type_Info_Procedure of the original command.
I found it useful to get access on the calling site for knowing if it takes parameters, and if it doesn't, just executing the command, if it does, put it into the input text so the user can put in the parameters. In general, it's nice to be able to link back to the procedure you originally annotated via jai's built-in type-info.